+++ src/web/assets/ckeditor/src/ckeditor5-craftcms.js
@@ -224,7 +224,7 @@ export const registerPackage = (pkg) => {
       const plugin = findPlugin(pluginName);
       if (!plugin) {
         console.warn(
-          `No plugin named ${pluginName} found in window.CKEditor5.`
+          `No plugin named ${pluginName} found in window.CKEditor5.`,
         );
         return;
       }
@@ -273,7 +273,7 @@ const trackChangesInSourceMode = function (editor) {
 
   sourceEditing.on('change:isSourceEditingMode', () => {
     const $sourceEditingContainer = $editorElement.find(
-      '.ck-source-editing-area'
+      '.ck-source-editing-area',
     );
 
     if (sourceEditing.isSourceEditingMode) {
@@ -303,10 +303,10 @@ export const create = async function (element, config) {
       ...pluginButtonMap
         .filter(
           ({buttons}) =>
-            !config.toolbar.some((button) => buttons.includes(button))
+            !config.toolbar.some((button) => buttons.includes(button)),
         )
         .map(({plugins}) => plugins)
-        .flat()
+        .flat(),
     );
   }
 
@@ -325,7 +325,7 @@ export const create = async function (element, config) {
 
   const editor = await ClassicEditor.create(
     element,
-    Object.assign({plugins}, config)
+    Object.assign({plugins}, config),
   );
 
   // Keep the source element updated with changes